Removing MysticEditor project.
Integrating the editor into the main mystic project.

Intended workflow:
(Mystic.csproj) The Editor class should hold all of the information and configuration for an editor, but the user (Sandbox.csproj) should be able to modify the editor as they make decisions inside of it.

There should be some configuration (docking) as well as things related to the project loading.

Also, I would like to add custom editor tools/windows via some sort of hooks in the future.