Adding getter/setter code generation with MSTRUCTs and MPROPERTYs
Starting to set the base for how objects will be represented in the world outside of an ECS.

Should I go with an approach similar to Unreal (see MObject,MStruct,MClass in this commit).
Should I go for a more Unity-style approach where a gameobject is ONLY a list of components, but those components have behaviors (MonoBehavior).

Regardless, it is clear that code generation via macros is doable.